+++ app/src/main/java/com/example/administrator/verwaltungstest/ProductListeActivity.java
@@ -7,7 +7,9 @@ import android.util.Log;
 import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
+import android.widget.Button;
 import android.widget.ListView;
+import android.widget.Toast;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -19,6 +21,7 @@ public class ProductListeActivity extends AppCompatActivity {
     private ListView productListView;
     private Datasource datasource;
     private Product slectedProduct;
+    private int selectFlag;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -26,6 +29,7 @@ public class ProductListeActivity extends AppCompatActivity {
         setContentView(R.layout.activity_product_liste);
         initializeKundeListView();
         datasource = new Datasource(this);
+        selectFlag = getIntent().getIntExtra("selectFlag",0);
     }
 
     @Override
@@ -42,12 +46,20 @@ public class ProductListeActivity extends AppCompatActivity {
         datasource.open();
         Log.d(TAG, "folgende Einträge sind in der DB vorhanden: ");
         showAllListEntries();
-        if (slectedProduct == null){
+        if (selectFlag == 123){
             findViewById(R.id.button_edit_product).setEnabled(false);
             findViewById(R.id.button_delete_product).setEnabled(false);
+            findViewById(R.id.button_edit_product).setVisibility(Button.INVISIBLE);
+            findViewById(R.id.button_delete_product).setVisibility(Button.INVISIBLE);
+            ((Button)findViewById(R.id.button_add_product)).setText("Ok");
         }else {
-            findViewById(R.id.button_edit_product).setEnabled(true);
-            findViewById(R.id.button_delete_product).setEnabled(true);
+            if (slectedProduct == null) {
+                findViewById(R.id.button_edit_product).setEnabled(false);
+                findViewById(R.id.button_delete_product).setEnabled(false);
+            } else {
+                findViewById(R.id.button_edit_product).setEnabled(true);
+                findViewById(R.id.button_delete_product).setEnabled(true);
+            }
         }
     }
 
@@ -81,9 +93,13 @@ public class ProductListeActivity extends AppCompatActivity {
     }
 
     public void productAnlegen(View view){
-        Intent intent = new Intent(this,ProductActivity.class);
-        intent.putExtra(getString(R.string.kunde_editmode),false);
-        startActivity(intent);
+        if (selectFlag == 123) {
+            selectProduct();
+        }else {
+            Intent intent = new Intent(this, ProductActivity.class);
+            intent.putExtra(getString(R.string.kunde_editmode), false);
+            startActivity(intent);
+        }
     }
 
     public void productEdit(View view){
@@ -95,7 +111,19 @@ public class ProductListeActivity extends AppCompatActivity {
 
     public void productDelete(View view){
         Product product = datasource.getProduct(slectedProduct.getId());
-        datasource.deleteProduct(product);
+        if(datasource.getAllLager_zu_Bestellungen_Product(product.getId()).size() == 0) {
+            datasource.deleteProduct(product);
+        }else{
+            Toast.makeText(ProductListeActivity.this, "Es sind noch Bestellungen für dieses Produkt vorhanden!", Toast.LENGTH_SHORT).show();
+        }
         showAllListEntries();
+    }
+
+    public void selectProduct(){
+        long id = slectedProduct.getId();
+        Intent intent = new Intent();
+        intent.putExtra("ProductId",id);
+        setResult(RESULT_OK,intent);
+        finish();
     }
 }
